[FEATURE] Non-breaking empty configuration
If the TypoScript is not (yet) included SettingsService throws an RuntimeException. If you are not already logged in to the backend, you won't be able to fix this: the exception is thrown before any chance to log into the backend :-/

All in all, throwing an exception is a bit exaggerated because of a missing configuration for an AuthService (that is maybe only used additionally) and not catching it.

This change will catch the exception and only report it as an error to the logging API.

Resolves: #39